requirements.txt
jinja2==3.1.6
supabase==2.10.0
python-docx==1.1.2
stripe==11.1.1
requests==2.32.3
